Move AssetCommand from assets to mission
AssetCommands can only exist in relation to a mission,
so move them to the mission app to avoid creating a
dependency loop between missions and assets
