+++ assets/admin.py
@@ -3,7 +3,7 @@ Expose assets to the admin interface
 """
 
 from django.contrib import admin
-from .models import Asset, AssetType, AssetCommand, AssetStatusValue
+from .models import Asset, AssetType, AssetStatusValue
 
 
 class AssetAdmin(admin.ModelAdmin):
@@ -20,13 +20,6 @@ class AssetTypeAdmin(admin.ModelAdmin):
     list_display = ['name', 'description']
 
 
-class AssetCommandAdmin(admin.ModelAdmin):
-    """
-    Allow admins to create asset commands
-    """
-    list_display = ['asset', 'command', 'reason', 'issued', 'issued_by', 'position']
-
-
 class AssetStatusValueAdmin(admin.ModelAdmin):
     """
     Allow admins to create asset status values
@@ -36,5 +29,4 @@ class AssetStatusValueAdmin(admin.ModelAdmin):
 
 admin.site.register(Asset, AssetAdmin)
 admin.site.register(AssetType, AssetTypeAdmin)
-admin.site.register(AssetCommand, AssetCommandAdmin)
 admin.site.register(AssetStatusValue, AssetStatusValueAdmin)
